Avoid rendering DayPicker when not visible
I was looking at some markup that we generate during server rendering
and noticed a whole bunch of markup (~32 KiB of HTML) belonging to
react-dates for elements that were not visible. With a couple of minor
modifications, we can avoid rendering most of these nodes until the
input is actually focused and we want to open the DayPicker.

This should give us a nice performance boost for both server-rendered
pages with this component and again on the client when rendering
happens.