Integrate with Happo.io
Happo.io is a screenshot testing service. It is used to find bugs,
visualize change and provide a visual log of your application's
components. It can be used to take screenshots in different browsers.

I've been using react-dates for a while when stress-testing happo. One
of the things that have been most tricky is that react-dates relies on
measuring elements in the DOM as part of rendering the UI. This is an
issue for default Happo, because it uses JSDOM to prerender components,
then collects the HTML and CSS and sends it to the happo.io to take
screenshots. JSDOM can't measure things. The pre-rendering phase is
important for speed and security, and is something that would be less
than trivial to bypass.  So, instead of bypassing pre-rendering I've
made it pre-render in a real browser. The `happo-plugin-puppeteer` npm
module injects a puppeteer-powered (uses Chrome) pre-renderer to Happo.
And by using that module, we're good to go for react-dates.

Another important module here is `happo-plugin-storybook`. This plugin
converts all storybook stories to happo examples. This means we can keep
using Storybook and we'll get screenshot testing "for free". I picked
the .storybook-css project and disabled a few decorators/addons in the
storybook config.

For now, we'll be taking screenshots in Chrome in two viewports:

- 900x600 ("desktop")
- 320x640 ("mobile")

Going forward, we may choose to add other targets/browsers as well. But
to begin with, I want to make sure things are working smoothly before we
expand.