Skip the install phase of the puppeteer dependency
Puppeteer is a dependency of happo-plugin-puppeteer. It is causing
installation errors on Node 4. By passing the
PUPPETEER_SKIP_CHROMIUM_DOWNLOAD flag I believe we can avoid the error.

I found out about this env flag in the Troubleshooting section in the
puppeteer docs:
https://github.com/GoogleChrome/puppeteer/blob/395c50624cdd82efa5516201d587b7e578944acf/docs/troubleshooting.md#troubleshooting

I still have to figure out how to get it installed. But this commit will
at least prevent the installation failure.