Install happo-plugin-puppeteer only when it's needed
Installing puppeteer (a dependency of happo-plugin-puppeteer) has proven
hard on Node 4. See for instance this Travis error log:

  > puppeteer@1.8.0 install /home/travis/build/airbnb/react-dates/node_modules/puppeteer
  > node install.js
  /home/travis/build/airbnb/react-dates/node_modules/puppeteer/install.js:74
    return Promise.all([...cleanupOldVersions, generateProtocolTypesIfNecessary(true /* updated */)]);
                        ^^^
  SyntaxError: Unexpected token ...

From https://travis-ci.org/airbnb/react-dates/jobs/435318847

Since we don't use happo in Node 4, we might as well delay installation
until we need the plugin, which is right before we run `happo`.

The downside of this approach is that running happo locally will be a
little slower, since the installation step is always run ahead of
invoking happo. And people might be confused about seeing a reference to
a package not listed in package.json.

This commit also includes a simplification to how the happo test script
is invoked by Travis. I had misunderstood how the matrix worked.