Add Happo API tokens as secrets to .travis.yml
To properly run happo for the project, we need to provide it with API
tokens. At first, I thought that an owner of the repo had to add these,
but if I understand things correctly, anyone is able to add keys.

https://docs.travis-ci.com/user/encryption-keys/

These keys will make Happo runs work for anyone with access to the repo.
They will not be available to forks/outside PRs (for security reasons).
Happo relies on a different auth mechanism for outside PRs, and for that
to work we need to install the Happo Github app. There are some
limitations to that auth mechanism however, and having the keys here for
members of the repo will help make things smoother.