Avoid some unnecessary re-rendering of CalendarDays
I have been investigating performance and noticed that when selecting a
date, all of the days for all of the months end up re-rendering. I
discovered that there are two causes for this currently:

1. All of the phrases are passed down to the CalendarDay components, and
   the phrases object can be different when selecting dates because of
   code like this, which toggles one translation back and forth:

   https://github.com/airbnb/react-dates/blob/888f6505/src/components/DayPickerRangeController.jsx#L501-L508

2. Similarly, all of the modifiers end up being different by reference
   because we end up creating new Sets for them every time, even if the
   sets are deeply equal.

To address the first issue, I decided to simplify the API of CalendarDay
a little, so that it takes an ariaLabel prop which is a simple string,
instead of a phrases prop. This allows us to prevent the changing of the
phrases object from automatically deoptimizing our leaf components.

And, to address the second issue, I introduced a simple cache for
modifier Sets. This allows the downstream components to perform a simple
referential equality check on them to avoid re-rendering, which we get
for free with PureComponent.

These two optimizations prevent many dates from re-rendering, namely
dates that are blocked and out of range. In storybook on my Macbook pro
(non-production React) this seems to bring down the time it takes to
update DayPickerRangeController when selecting dates from ~150ms to
~110ms. I think this will have a positive effect on some other
operations or other types of datepickers, but I didn't profile them.

It is worth noting that there is still a lot of re-rendering happening,
but that is because the value of ariaLabel ends up changing while you
select dates from:

> Choose Thursday, May 30, 2019 as your check-in date. It’s available.

to:

> Choose Thursday, May 30, 2019 as your check-out date. It’s available.

This seems important for accessibility, and I don't have any good ideas
at this moment of how to avoid this yet.