Convert CalendarDay's aria-label to aria-labelledby
I've been working to improve performance of react-dates, and I noticed
that on many interactions every single date ends up being re-rendered.
In my previous commit, I partly fixed this by avoiding passing the
always-new phrases object down to CalendarDay, and to memoize modifiers
so they also don't deoptimize the pure components at the edges.

However, this still left one phrase that always changes when
selecting start and end dates which still caused most dates to be
re-rendered.

To avoid this, instead of passing an aria-label to CalendarDay, which
needs to change when switching from selecting a start date to selecting
an end date, we can render a div that contains the phrase for the
currently focused date, and pass a reference to it to be used by
aria-labelledby.

The downside here is that since ids need to be unique per page and there
is no reliable way to auto-generate an id on the server that will be
consistent on the client, I had to add required id props in a number of
places. For folks using components that already require startDateId, we
can generate a new id based on that value, so the friction there should
be minimal.

Testing this using VoiceOver on Safari and the results seem to be just
what we want.

Profiling this in storybook (non-production React) on my Macbook pro
brings the time it takes to update DayPickerRangeController from ~110ms
after my previous commit to ~75ms. (Before both of these commits, we
were looking at ~150ms).

I think there might still be some opportunity in preventing unnecessary
re-renders of CalendarMonth, but most of the remaining opportunity looks
like it might be in optimizing how modifiers are built in places like
DayPickerRangeController#componentWillReceiveProps.