Remove "Show info" button from Happo snapshots
The info addon is showing a blue button at the top for all stories. We
don't want that to end up in the happo screenshots. I tried to find a
more elegant way than wrapping the `withInfo` function, but in the end
the wrapper made most sense. There is of course a risk people will
forget to use the wrapper when they add new stories, but that wouldn't
be the end of the world.