Add a 200ms delay for Happo
While testing the happo run, I've noticed that a number of story
screenshots are cut off. Happo is quite eager at taking the screenshot,
and I think what's happening is that the component isn't fully done
rendering at the time the screenshot is taken.

Using a delay isn't ideal as A) there is still a risk for cutoff
screenshots, and B) it increases the total run time by 200ms times the
number of stories. In a mature open source project like this, I don't
expect the test suite to be run more than perhaps a few times per day
however, so the performance thing might not be terrible after all. I'll
continue investigating whether we can do better. One idea I had was that
stories themselves can call a function when they're done. Something like

  import { done } from 'happo-plugin-storybook/register';

  storiesOf('DayPicker', module)
    .add('default', withInfo()(() => (
      <DayPicker onDone={done} />
    )), { happo: { usesCallback: true } })

I don't yet know how to implement that behavior though, both in the
component itself and on the happo-plugin-storybook side.