Add core-js@3 as a dev dependency
After installing happo.io in this project after the upgrade to Storybook
v5, I started running in to problems with the wrong core-js version
being used. @storybook/core is depending on core-js@3 but the version
hoisted at the root of node_modules was v2. Which should be fine, any
@storybook/core code should pick the local version
(the one found under node_modules/@storybook/core/node_modules). But
that didn't seem to work in all cases, and I saw things like

  ERROR in ./.storybook/config.js
  Module not found: Error: Can't resolve 'core-js/modules/es.array.for-each' in '/indio-ui/.storybook'
   @ ./.storybook/config.js 1:0-43
   @ multi ./node_modules/@storybook/core/dist/server/common/polyfills.js ./node_modules/@storybook/core/dist/server/preview/globals.js ./.storybook/config.js ./node_modules/webpack-hot-middleware/client.js?reload=true

The happo.io library depends on core-js@2, and it was after installing
happo.io that v2 started getting hoisted into the root. It seems
Storybook is always using the root-most version, not sure why. This
issue hints at many others running into the same issue, or something
similar:
https://github.com/storybookjs/storybook/issues/6204

There are many copies of core-js in our dependency tree. `karma` depends
on core-js@2, `react-addons-shallow-compare` depends on core-js@1 (via
`fbjs`).

Adding a direct dependency on core-js resolves the issues. Once there is
a proper solution for
https://github.com/storybookjs/storybook/issues/6204, we can most likely
remove this dependency.