Update screenreader copy to be more generic, allow custom aria labels
Original "default phrases" for this package were very airbnb specific,
so we removed references to hotels/check-ins.

The default behavior on the DatePickerInput component uses the input's
placeholder as the aria-label, and it isn't customizable beyond that.
Now, we've implemented a default, generic aria-label, or a custom one
can be applied to the component as necessary.

Default behavior also removed the tabindex from custom navigation
buttons. We moved out the logic assigning the tabindex so it would
be applied on custom buttons as well.

Resolves DESLANG-2577