Remove Moment;  replace with driver interface and Moment support.
STATUS: WIP

This PR opens the door to supporting multiple date libraries.

Instead of calling moment directly within each component, we instead
create a driver which wraps all functions we need from moment.  This
driver supports (theoretically) multiple runtimes;  we can use moment or
swap out for Luxon/date-fns/js-joda as we wish.

The implementation of the driver is similar to a go interface.  A driver
must implement all of the calls that we use throughout our components to
fulfil the interface.

The driver also assumes that any function call is stateless;  calling
`driver.add(date, { days: 2 })` must return the new date without
mutating the argument.

As most date functions accept the library's date instance as the first
argument we also provide a utility to curry these functions w/ partial
application:

```js

import { driver, chain } from '../src/drivers/drivers';

chain(date).add({ days: 1}).startOf('day').value();
```

This is a WIP PR.  I'm not particulary happy with the way that drivers
are defined, with the way that they're instantiated (via mutating), and
there are a few errors that I need to get to.  But it's the direction
I'm heading in after wanting to use this awesome datepicker with Luxon.

Thoughts appreciated as I work on it :)