Fix the focusing on CalendarDay when isFocused and tabIndex = 0.
This was previously only focusing in componentDidUpdate so it would have to wait until something changed before setting the focus on the CalendarDay.  Additionally there is a timing issue where the buttonRef exists but calling focus on it does not work if done during the first tick.  Using setTimeout (vs raf) delays this call just long enough that it is able to fire.