Resolve issue 1023
When a date is invalid, it is hard to tell if it is invalid due to an empty string, where the user has not entered a date, or if the user entered an invalid date.

Currently the only way to determine if the date is invalid or just empty, is to query the DOM for the specific tags to read their values.

This change will pass a new argument to the onChangeHandler that will include the date string that was entered into the input.