[Fix] `DateRangePicker`, `DateRangePickerInput`: Fix keyboard navigation
Keyboard navigation is broken for the date range picker. The following
bugs exist:

When passing through the component with tab-navigation, after exiting
the end-date input the dropdown does not close. When attempting to enter
values into the end-date dropdown, users would have to tab backwards
back to the dropdown to enter values. The simplest and clearest solution
that came to be is simply to put the day picker after the currently
active input. In this way, the dropdown always appears next in the
tab-order to the input the user has selected. The existing blur handler
in the dropdown means that tabbing through the component will leave the
dropdown closed at the end.

The react keys ensure that the element is just moved over in the DOM and
so there is no flickering.

Fixes #1809

Co-authored-by: Greg O'Connor <greg.oconnor@airbnb.com>
Co-authored-by: Daniel M <daniel.magarreiro@airbnb.com>